adding enterprise API client to hit customer endpoint
