+++ enterprise_catalog/apps/api_client/base_oauth.py
@@ -1,0 +1,22 @@
+from django.conf import settings
+from edx_rest_api_client.client import OAuthAPIClient
+
+
+class BaseOAuthClient:
+    """
+    Base class for OAuth API clients.
+    """
+    def __init__(self):
+        self.client = OAuthAPIClient(
+            settings.SOCIAL_AUTH_EDX_OAUTH2_URL_ROOT.strip('/'),
+            self.oauth2_client_id,
+            self.oauth2_client_secret
+        )
+
+    @property
+    def oauth2_client_id(self):
+        return settings.BACKEND_SERVICE_EDX_OAUTH2_KEY
+
+    @property
+    def oauth2_client_secret(self):
+        return settings.BACKEND_SERVICE_EDX_OAUTH2_SECRET
